Feat : 동시성 문제 해결 (concurrentHashMap, concurrentLinkedqueue)

+++ src/main/java/caugarde/vote/service/StudentVoteService.java
@@ -1,20 +1,29 @@
 package caugarde.vote.service;
 
-import caugarde.vote.model.constant.CustomOAuthUser;
+import caugarde.vote.model.constant.VoteTask;
 import caugarde.vote.model.dto.response.StudentVoteResponseDTO;
+import caugarde.vote.model.entity.Category;
 import caugarde.vote.model.entity.Student;
 import caugarde.vote.model.entity.StudentVote;
 import caugarde.vote.model.entity.Vote;
+import caugarde.vote.repository.jpa.CategoryRepository;
 import caugarde.vote.repository.jpa.StudentRepository;
 import caugarde.vote.repository.jpa.StudentVoteRepository;
 import caugarde.vote.repository.jpa.VoteRepository;
+import jakarta.annotation.PreDestroy;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedQueue;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,9 +33,85 @@ public class StudentVoteService {
 
     private final StudentVoteRepository studentVoteRepository;
     private final VoteRepository voteRepository;
-    private final StudentRepository studentRepository;
+    private final CategoryRepository categoryRepository;
 
-    public Boolean save(UUID id, Student student) {
+    private final ExecutorService executorService = Executors.newFixedThreadPool(10);
+    private final ConcurrentLinkedQueue<VoteTask> voteSaveQueue = new ConcurrentLinkedQueue<>();
+
+    private final ConcurrentHashMap<UUID, VoteTask> cancelTaskMap = new ConcurrentHashMap<>();
+
+    // 투표 요청 처리 메서드
+    public void processVote(UUID voteId, Student student, Category category, Consumer<String> messageHandler, Consumer<StudentVoteResponseDTO.CreateAttendanceNumber> countHandler) {
+
+        voteSaveQueue.add(new VoteTask(voteId, student,category));
+
+        // 비동기적으로 큐에서 작업을 처리
+        executorService.submit(() -> processVoteQueue(messageHandler, countHandler));
+
+    }
+
+    private void processVoteQueue(Consumer<String> messageHandler, Consumer<StudentVoteResponseDTO.CreateAttendanceNumber> countHandler) {
+
+        while (!voteSaveQueue.isEmpty()) {
+            VoteTask task = voteSaveQueue.poll();
+            boolean isSaved = save(task.getVotePk(), task.getStudent(),task.getCategory());
+
+            // 메시지와 카운트 업데이트
+            StudentVoteResponseDTO.CreateMessage resultMessageDTO = new StudentVoteResponseDTO.CreateMessage(isSaved ? "투표가 완료되었습니다." : "이미 투표하셨습니다.");
+            messageHandler.accept(resultMessageDTO.message());
+
+            StudentVoteResponseDTO.CreateAttendanceNumber voteCountDTO = new StudentVoteResponseDTO.CreateAttendanceNumber(
+                    task.getVotePk(), countAllByVotePk(task.getVotePk()), voteRepository.findById(task.getVotePk()).get().getLimitPeople()
+            );
+            countHandler.accept(voteCountDTO);
+
+        }
+
+    }
+
+
+    public void cancelVote(UUID voteId, Student student, Consumer<String> messageHandler, Consumer<StudentVoteResponseDTO.CreateAttendanceNumber> countHandler) {
+        // 작업을 Map에 추가
+
+        UUID studentVoteId = studentVoteRepository.findByVoteAndStudent(voteRepository.findById(voteId).get(),student).get().getStudentVotePk();
+        cancelTaskMap.put(studentVoteId, new VoteTask(voteId, student));
+
+        // 비동기적으로 Map에서 작업을 처리
+        executorService.submit(() -> cancelVoteTasks(messageHandler, countHandler));
+    }
+
+    private void cancelVoteTasks(Consumer<String> messageHandler, Consumer<StudentVoteResponseDTO.CreateAttendanceNumber> countHandler) {
+        // 맵에서 작업을 처리할 수 있을 때까지 반복
+        List<UUID> tasksToProcess = new ArrayList<>(cancelTaskMap.keySet());
+
+        for (UUID studentVotePk : tasksToProcess) {
+            VoteTask task = cancelTaskMap.get(studentVotePk);
+            if (task != null) {
+                // 처리할 작업 수행
+                Boolean isDeleted = deleteById(studentVotePk);
+
+                // 메시지와 카운트 업데이트
+                StudentVoteResponseDTO.DeleteMessage resultMessageDTO = new StudentVoteResponseDTO.DeleteMessage(isDeleted ? "투표가 취소되었습니다." : "투표 내역이 없습니다.");
+                messageHandler.accept(resultMessageDTO.message());
+
+                StudentVoteResponseDTO.CreateAttendanceNumber voteCountDTO = new StudentVoteResponseDTO.CreateAttendanceNumber(
+                        task.getVotePk(), countAllByVotePk(task.getVotePk()), voteRepository.findById(task.getVotePk()).get().getLimitPeople()
+                );
+                countHandler.accept(voteCountDTO);
+
+                // 처리된 작업을 맵에서 제거
+                cancelTaskMap.remove(studentVotePk);
+            }
+        }
+    }
+
+
+    @PreDestroy
+    public void shutdown() {
+        executorService.shutdown();
+    }
+
+    public Boolean save(UUID id, Student student,Category category) {
 
         Vote vote = voteRepository.findById(id).orElseThrow(EntityNotFoundException::new);
         if(studentVoteRepository.findByVoteAndStudent(vote, student).isPresent()) {
@@ -35,15 +120,24 @@ public class StudentVoteService {
             StudentVote studentVote = StudentVote.builder()
                     .studentVotePk(UUID.randomUUID())
                     .student(student)
-                    .vote(voteRepository.findById(id).orElse(null))
+                    .vote(vote)
+                    .category(category)
                     .build();
 
             studentVoteRepository.save(studentVote);
-            student.getStudentVotes().add(studentVote);
-            studentRepository.save(student);
             return true;
         }
+    }
 
+    public List<StudentVote> getVoteInfo(UUID votePk){
+        Vote vote = voteRepository.findById(votePk).orElse(null);
+        return studentVoteRepository.findByVote(vote);
+    }
+
+    public List<StudentVoteResponseDTO> convertToSVDTO(List<StudentVote> studentVoteList) {
+        return studentVoteList.stream()
+                .map(StudentVoteResponseDTO::new)
+                .collect(Collectors.toList());
     }
 
     public List<Student> getStudentsByVoteId(UUID id) {
@@ -56,26 +150,13 @@ public class StudentVoteService {
                 .collect(Collectors.toList());
     }
 
-    public List<StudentVoteResponseDTO> studentsToDTOs(List<Student> students) {
-        return students.stream().map(StudentVoteResponseDTO::new).collect(Collectors.toList());
-    }
-
-    public void deleteByVoteAndStudent(UUID id, CustomOAuthUser user) {
-        Student student = studentRepository.findById(user.getId()).orElse(null);
-        Vote vote = voteRepository.findById(id).orElse(null);
-        studentVoteRepository.deleteByVoteAndStudent(vote, student);
-    }
-
-    public List<StudentVote> findAll() {
-        return studentVoteRepository.findAll();
-    }
-
-    public StudentVote findById(UUID id) {
-        return studentVoteRepository.findById(id).orElse(null);
-    }
-
-    public void deleteById(UUID id) {
-        studentVoteRepository.deleteById(id);
+    public boolean deleteById(UUID id) {
+        if (studentVoteRepository.findById(id).isPresent()){
+            studentVoteRepository.deleteById(id);
+            return true;
+        }
+        else
+            return false;
     }
 
     public void deleteByVote(UUID id) {
@@ -88,8 +169,11 @@ public class StudentVoteService {
     }
 
     public int countByVote(Vote vote){
-        return studentVoteRepository.findByVote(vote).size();
+        return (int) studentVoteRepository.countByVote(vote);
     }
 
+    public int countAllByVotePk(UUID id){
+        return countByVote(voteRepository.findById(id).get());
+    }
 
 }
